% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accessor_functions.R
\name{Accessors}
\alias{Accessors}
\alias{get_intercept}
\alias{get_y}
\alias{get_size}
\alias{get_hyp}
\alias{get_lambda}
\alias{get_psi}
\alias{get_se}
\alias{get_kernels}
\alias{get_kern_matrix}
\alias{get_mse}
\alias{get_estl}
\alias{get_method}
\alias{get_convergence}
\alias{get_niter}
\alias{get_time}
\title{Accessor functions for \code{ipriorMod} objects.}
\usage{
get_intercept(object)

get_y(object)

get_size(object, units = "kB", standard = "SI")

get_hyp(object)

get_lambda(object)

get_psi(object)

get_se(object)

get_kernels(object)

get_kern_matrix(object, theta = NULL, xstar = list(NULL))

get_mse(object)

get_estl(object)

get_method(object)

get_convergence(object)

get_niter(object)

get_time(object)
}
\arguments{
\item{object}{An \code{ipriorMod} object.}

\item{units}{Units for object size.}

\item{standard}{Standard for object size.}

\item{theta}{(Optional) Value of hyperparameters to evaluate the kernel
matrix.}

\item{xstar}{(Optional) If not supplied, then a square, symmetric kernel
matrix is returned using the data as input points. Otherwise, the kernel
matrix is evaluated with respect to this set of data as well. It must be a
list of vectors/matrices with similar dimensions to the original data.}
}
\description{
Accessor functions for \code{ipriorMod} objects.
}
