% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{as.time}
\alias{as.time}
\title{Convert \code{difftime} class into \code{time} class}
\usage{
as.time(x)
}
\arguments{
\item{x}{A \code{difftime} object.}
}
\value{
A \code{time} object which contains the time difference and units.
}
\description{
Convert \code{difftime} class into \code{time} class
}
