% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betas.R
\name{betas}
\alias{betas}
\title{Function which returns the Variance of Y and log-likelihood value.}
\usage{
betas(object)
}
\arguments{
\item{x}{Object of class "iprior"}

\item{alpha}{(Optional) specify intercept value.}

\item{lambda}{(Optional) specify lambda value.}

\item{psi}{(Optional) specify psi value.}
}
\value{
List of length 2.
}
\description{
\code{betas} returns the betas.
}
\details{
Test
}
\examples{
mod.iprior <- iprior(stack.loss ~ ., data=stackloss)
betas(mod.iprior)
}

