% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{check_theta}
\alias{check_theta}
\title{Check the structure of the hyperparameters of an I-prior model}
\usage{
check_theta(object)
}
\arguments{
\item{object}{An \code{ipriorMod} object or an \code{ipriorKernel} object.}
}
\value{
A printout of the structure of the hyperparameters.
}
\description{
Check the structure of the hyperparameters of an I-prior model
}
