% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datfbm}
\alias{datfbm}
\title{Simulated data to illustrate one-dimensional smoothing}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{\code{x}}{x}
  \item{\code{y}}{y}
}

#' @examples
data(simdat)
str(simdat)}
\usage{
datfbm
}
\description{
Description.
}
\details{
Details.
}
\keyword{datasets}

