% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{decimal_place}
\alias{decimal_place}
\alias{dec_plac}
\title{Cut a numeric vector to a certain number of decimal places}
\usage{
decimal_place(x, k = 2)

dec_plac(x, k = 2)
}
\arguments{
\item{x}{A numeric vector.}

\item{k}{The number of decimal places.}
}
\value{
A character vector with the correct number of decimal places.
}
\description{
Cut a numeric vector to a certain number of decimal places
}
\examples{
decimal_place(pi, 3)
decimal_place(c(exp(1), pi, sqrt(2)), 4)

}
