% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EigenCpp}
\alias{EigenCpp}
\title{Eigen decomposition of a matrix in C++}
\usage{
EigenCpp(M)
}
\arguments{
\item{M}{A symmetric, positive-definite matrix}
}
\description{
Returns the eigenvalues and eigenvectors of a matrix \code{M}.
}
\details{
A fast implementation of \code{eigen()} for symmetric,
positive-definite matrices. This helps speed up the I-prior EM algorithm.
}

