% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastSquare}
\alias{FastSquare}
\title{Multiplying a symmetric matrix by itself in C++}
\usage{
FastSquare(AA)
}
\arguments{
\item{AA}{A symmetric matrix}
}
\description{
Returns the square of a symmetric matrix \code{AA}.
}

