% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastVdiag}
\alias{FastVdiag}
\title{Computing a quadratic matrix form in C++}
\usage{
FastVdiag(X, y)
}
\arguments{
\item{X}{A symmetric, square matrix of dimension \code{n} by \code{n}}

\item{y}{A vector of length \code{n}}
}

