% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iprior_helper.R
\name{get_kern_matrix}
\alias{get_kern_matrix}
\title{Extract the kernel matrix from I-prior models}
\usage{
get_kern_matrix(object, theta = NULL, xstar = list(NULL))
}
\arguments{
\item{object}{An \code{ipriorMod} or \code{ipriorKernel2} object.}

\item{theta}{(Optional) Value of hyperparameters to evaluate the kernel
matrix.}

\item{xstar}{(Optional) If not supplied, then a square, symmetric kernel
matrix is returned using the data as input points. Otherwise, the kernel
matrix is evaluated with respect to this set of data as well. It must be a
list of vectors/matrices with similar dimensions to the original data.}
}
\value{
A kernel matrix.
}
\description{
Extract the kernel matrix from I-prior models
}
