% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{gg_colour_hue}
\alias{gg_colour_hue}
\alias{gg_color_hue}
\alias{gg_col_hue}
\alias{ipriorColPal}
\alias{ggColPal}
\title{Emulate \code{ggplot2} default colour palette}
\usage{
gg_colour_hue(x, h = c(0, 360) + 15, c = 100, l = 65)

gg_color_hue(x, h = c(0, 360) + 15, c = 100, l = 65)

gg_col_hue(x, h = c(0, 360) + 15, c = 100, l = 65)

ipriorColPal(x)

ggColPal(x)
}
\arguments{
\item{x}{The number of colours required.}

\item{h}{Range of hues to use, in [0, 360].}

\item{c}{Chroma (intensity of colour), maximum value varies depending on
combination of hue and luminance.}

\item{l}{Luminance (lightness), in [0, 100].}
}
\description{
Emulate \code{ggplot2} default colour palette. \code{ipriorColPal} and
\code{ggColPal} are DEPRECATED.
}
\details{
This is the default colour scale for categorical variables in \code{ggplot2}.
It maps each level to an evenly spaced hue on the colour wheel. It does not
generate colour-blind safe palettes.

\code{ipriorColPal()} used to provide the colour palette for the
\code{iprior} package, but this has been changed \code{ggplot2}'s colour
palette instead.
}
