% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hlam.R
\name{Hlam}
\alias{Hlam}
\alias{Hlam.ipriorKernel}
\alias{Hlam.ipriorMod}
\title{Extract the scaled kernel matrix}
\usage{
Hlam(object, lambda = NULL)

\method{Hlam}{ipriorMod}(object, lambda = NULL)

\method{Hlam}{ipriorKernel}(object, lambda = NULL)
}
\arguments{
\item{object}{An object of class \code{ipriorMod} or \code{ipriorKernel}.}

\item{lambda}{Values of the scale parameters.}
}
\description{
Extract the scaled kernel matrix \deqn{\mathbf H_\lambda}.
}
\details{
For \code{ipriorKernel} objects, random values for \code{lambda} are used.
}

