% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{hsbsmall}
\alias{hsbsmall}
\title{High school and beyond dataset}
\format{A data frame of 661 observations on 3 variables. \describe{
  \item{\code{mathach}}{Math achievement.} \item{\code{ses}}{Socio-Economic
  status.} \item{\code{schoolid}}{Categorical variable indicating the school
  the student went to. Treated as \code{factor}.} }}
\usage{
hsbsmall
}
\description{
Smaller subset of \code{hsb}.
}
\details{
A random subset of size 16 out of the original 160 groups.
}
\examples{
data(hsbsmall)
str(hsbsmall)
}
\keyword{datasets}
