\name{iprior-package}
\alias{iprior-package}
\alias{iprior}
\docType{package}
\title{
\packageTitle{iprior}
}
\description{
\packageDescription{iprior}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{iprior}
\packageIndices{iprior}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{iprior}

Maintainer: \packageMaintainer{iprior}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
