% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iprior-package.R
\docType{package}
\name{iprior_package}
\alias{iprior_package}
\alias{iprior_package-package}
\title{\code{iprior}: Regression using priors with Fisher information covariance
kernels.}
\description{
The \code{iprior} package provides methods to perform and analyse I-prior
regression models. Estimation is done either via direct optimisation of the
log-likelihood or an EM algorithm.
}
\section{Author(s)}{


  \strong{Maintainer}: Haziq Jamil

  Contributors: \itemize{\item{Wicher Bergsma}}
}

\seealso{
\url{http://phd.haziqj.ml/}
}
