% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipriorNystrom_plots.R
\name{iprior_plot}
\alias{iprior_plot}
\alias{plot_fitted}
\alias{plot_multilevel}
\title{New plots}
\usage{
plot_fitted(object)

plot_multilevel(object)
}
\arguments{
\item{object}{an iprior object}
}
\description{
New plots
}
