% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{ipriorColPal}
\alias{ipriorColPal}
\title{Colour palette for \code{iprior} plots}
\usage{
ipriorColPal(x)
}
\arguments{
\item{x}{(optional) A vector of maximum length 16.}
}
\value{
The colour palette indexed by \code{x}.
}
\description{
This is the colour palette used by the \code{iprior} package. It is based off
\code{RColorBrewer::brewer.pal}'s Set 1, Set 2 and Dark 2 palettes.
}
