% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{ipriorloglik}
\alias{ipriorloglik}
\title{Function which returns the Variance of Y and log-likelihood value.}
\usage{
ipriorloglik(x, alpha = NULL, lambda = NULL, psi = NULL)
}
\arguments{
\item{x}{Object of class "iprior"}

\item{alpha}{(Optional) specify intercept value.}

\item{lambda}{(Optional) specify lambda value.}

\item{psi}{(Optional) specify psi value.}
}
\value{
List of length 2.
}
\description{
\code{ipriorloglik} returns the the variance of Y and the log-likelihood value.
}
\details{
Test
}
\examples{
mod.iprior <- iprior(stack.loss ~ ., data=stackloss)
ipriorloglik(mod.iprior)
}

