% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipriorNystrom.R
\name{ipriorNystrom}
\alias{ipriorNystrom}
\alias{ipriorNystrom.default}
\alias{ipriorNystrom.formula}
\title{Nystrom method to fit I-prior models}
\usage{
\method{ipriorNystrom}{default}(y, ..., size, seed = NULL,
  lambda.init = NULL, psi.init = NULL, model = list())

\method{ipriorNystrom}{formula}(formula, data, size, seed = NULL,
  lambda.init = NULL, psi.init = NULL, model = list(), ...)
}
\arguments{
\item{y}{Response variable}

\item{...}{Covariates}

\item{size}{Size of Nystrom sample}

\item{seed}{Choose a random seed, defaults to NULL}

\item{lambda.init}{NULL}

\item{psi.init}{NULL}

\item{model}{Model type}

\item{formula}{formula}

\item{data}{data}
}
\description{
Nystrom method to fit I-prior models
}
