% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{is.iprior_x}
\alias{is.iprior_x}
\alias{is.ipriorMod}
\alias{is.ipriorKernel}
\alias{is.nystrom}
\title{Test \code{iprior} objects}
\usage{
is.ipriorMod(x)

is.ipriorKernel(x)

is.nystrom(x)
}
\arguments{
\item{x}{An \code{ipriorMod} or \code{ipriorKernel} object.}
}
\value{
Logical.
}
\description{
Test whether an object is an \code{ipriorMod}, \code{ipriorKernel}, or either
object with Nystrom method enabled.
}
