% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{is.ipriorMod2}
\alias{is.ipriorMod2}
\title{Test for \code{iprior} objects}
\usage{
is.ipriorMod2(x)
}
\arguments{
\item{x}{An object.}
}
\description{
Checks whether an object is an \code{iprior} fitted model (i.e. an
\code{ipriorMod} object), or an object ready for an \code{iprior} fit (i.e.
an \code{ipriorKernel} object).
}
