% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{is.kern_x}
\alias{is.kern_x}
\alias{is.kern_linear}
\alias{is.kern_canonical}
\alias{is.kern_fbm}
\alias{is.kern_pearson}
\alias{is.kern_se}
\alias{is.kern_poly}
\title{Test kernel attributes}
\usage{
is.kern_linear(x)

is.kern_canonical(x)

is.kern_fbm(x)

is.kern_pearson(x)

is.kern_se(x)

is.kern_poly(x)
}
\arguments{
\item{x}{An \code{ipriorMod} object, \code{ipriorKernel} object, a kernel
matrix generated from one of the \code{kern_x()} functions, or even simply
just a character vector.}
}
\value{
Logical.
}
\description{
Test whether an object uses a specific type of kernel.
}
