% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel_loader.R
\name{kernL}
\alias{kernL}
\alias{kernL.formula}
\title{Load the kernel matrices for \code{iprior} fit}
\usage{
kernL(y, ..., model = list())

\method{kernL}{formula}(formula, data, model = list(), ...)
}
\arguments{
\item{y}{Vector of response variables.}

\item{...}{Only for when fitting using non-formula, enter the variables
(vectors or matrices) separated by commas. No other options applicable
here.}

\item{model}{(optional) List of model options. Not used for
\code{ipriorKernel} or \code{ipriorModel} objects.}

\item{formula}{The formula to fit when using formula interface.}

\item{data}{Data frame containing variables when using formula interface.}
}
\value{
an object of class ipriorKernel.
}
\description{
Description.
}
\details{
Details.
}
\examples{
str(ToothGrowth)
mod <- kernL(y = ToothGrowth$len, supp = ToothGrowth$supp,
             dose = ToothGrowth$dose, model = list(interactions="1:2"))
summary(mod)
\dontrun{mod <- kernL(len ~ supp * dose, data = ToothGrowth)  # also accepts
formula}

}

