% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{loglik}
\alias{loglik}
\title{Function which returns the log-likelihood value of an \code{iprior} or \code{ipriorKernel} object.}
\usage{
loglik(object, ...)
}
\arguments{
\item{x}{Object of class "iprior"}

\item{alpha}{(Optional) specify intercept value.}

\item{lambda}{(Optional) specify lambda value.}

\item{psi}{(Optional) specify psi value.}
}
\value{
List of length 2.
}
\description{
\code{loglik} returns the log-likelihood value.
}
\details{
\code{deviance} returns twice the negative log-likelihood value.

Test
}
\examples{
mod.iprior <- iprior(stack.loss ~ ., data=stackloss)
loglik(mod.iprior)
}

