% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{mod1.cv}
\alias{mod1.cv}
\alias{mod4.cv}
\title{Tecator CV}
\format{Results from iprior_cv cross validation experiment.}
\usage{
mod1.cv

mod4.cv
}
\description{
Tecator CV
}
\keyword{datasets}
