% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot.ipriorMod}
\alias{plot.ipriorMod}
\alias{plot_predict}
\alias{plot_fitted}
\alias{plot_iter}
\title{Plots for I-prior models}
\usage{
\method{plot}{ipriorMod}(x, ...)

plot_predict(x)

plot_fitted(x, X.var = 1, ci = TRUE)

plot_iter(x, niter.plot = NULL, lab.pos = c("up", "down"))
}
\arguments{
\item{x}{An \code{ipriorMod} object.}

\item{...}{Not used}

\item{X.var}{The index of the X variable to plot.}

\item{ci}{Logical. Plot the confidence intervals? Defaults to \code{TRUE}.}

\item{niter.plot}{(Optional) Vector of length at most two, indicating the
start and end points of the iterations to plot.}

\item{lab.pos}{Adjust the position of the log-likelihood label.}
}
\description{
There are three types of plots that are currently written in the package:
\describe{ \item{\code{plot_fitted}}{Plot the fitted regression line with
credibility bands.} \item{\code{plot_predict}}{Plot residuals against fitted
values.} \item{\code{plot_iter}}{Plot the progression of the log-likelihood
value over time.} } The S3 method \code{plot} for class \code{ipriorMod}
currently returns \code{plot_fitted}.
}
