% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{pollution}
\alias{pollution}
\title{Air pollution and mortality}
\format{A data frame of 16 observations on 16 variables.
\describe{
  \item{\code{Mortality}}{Total age-adjusted mortality rate per 100,000.}
  \item{\code{Rain}}{Mean annual precipitation in inches.}
  \item{\code{Humid}}{Mean annual precipitation in inches.}
  \item{\code{JanTemp}}{Mean annual precipitation in inches.}
  \item{\code{JulTemp}}{Mean annual precipitation in inches.}
  \item{\code{Over65}}{Mean annual precipitation in inches.}
  \item{\code{Popn}}{Mean annual precipitation in inches.}
  \item{\code{Educ}}{Mean annual precipitation in inches.}
  \item{\code{Hous}}{Mean annual precipitation in inches.}
  \item{\code{Dens}}{Mean annual precipitation in inches.}
  \item{\code{NonW}}{Mean annual precipitation in inches.}
  \item{\code{WhiteCol}}{Mean annual precipitation in inches.}
  \item{\code{Poor}}{Mean annual precipitation in inches.}
  \item{\code{HC}}{Mean annual precipitation in inches.}
  \item{\code{NOx}}{Mean annual precipitation in inches.}
  \item{\code{SO2}}{Mean annual precipitation in inches.}
}}
\usage{
pollution
}
\description{
Data on the relation between weather, socioeconomic, and air pollution
variables and mortality rates in 60 Standard Metropolitan Statistical Areas
(SMSAs) of the USA, for the years 1959-1961.
}
\examples{
data(pollution)
str(pollution)
}
\references{
McDonald, G. C. and Schwing, R. C. (1973). Instabilities of
  regression estimates relating air pollution to mortality.
  \emph{Technometrics}, 15(3):463-481.
}
\keyword{datasets}
