% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.iprior}
\alias{predict.iprior}
\title{Predict for iprior models.}
\usage{
\method{predict}{iprior}(object, newdata = list(), ...)
}
\arguments{
\item{object}{Objects of class \code{iprior}.}

\item{newdata}{(Optional) A data frame in which to look for variables with
which to predict. If omitted, the fitted values are used.}
}
\description{
Predict for iprior models.
}
\details{
Predict  for iprior models.
}
\examples{
\donttest{mod <- iprior(stack.loss ~ ., data=stackloss)}
\donttest{predict(mod, newdata = data.frame(c(1,2,3)))}

}

