% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress}
\alias{progress}
\title{Progress for fitted \code{ipriorMod} object}
\usage{
progress(x, interval = c("auto", "all", "input any number"))
}
\arguments{
\item{x}{x}

\item{interval}{interval}
}
\description{
Description.
}
\details{
Details.
}

