% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{sigma}
\alias{sigma}
\alias{sigma.ipriorMod}
\title{Obtain the standard deviation of the residuals 'sigma'}
\usage{
\method{sigma}{ipriorMod}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ipriorMod}.}

\item{...}{This is not used here.}
}
\description{
Extract the standard deviation of the residuals. For I-prior models, this is
\code{sigma = 1 / sqrt(psi)}.
}
\details{
This basically obtains \code{object$sigma}. For \code{R (>= 3.3.0)} then
\code{sigma} is an S3 method with the default method coming from the
\code{stats} package.
}

