% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{simdat}
\alias{simdat}
\title{Random slopes model simulated data}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{x}{x}
  \item{y}{y}
  \item{grp}{grp}
}

#' @examples
data(simdat)
str(simdat)}
\usage{
simdat
}
\description{
Description.
}
\details{
Details.
}
\keyword{datasets}

