% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope.R
\name{slope}
\alias{slope}
\title{Function which returns the Variance of Y and log-likelihood value.}
\usage{
slope(ipriorMod)
}
\arguments{
\item{ipriorMod}{Object of class "ipriorMod"}
}
\value{
What.
}
\description{
\code{slope} returns the slopes.
}
\details{
Details.
}
\examples{
mod.iprior <- iprior(stack.loss ~ Air.Flow, data=stackloss)
slope(mod.iprior)

}

