% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iprior.R
\name{summary.iprior}
\alias{summary.iprior}
\title{Summary screen for iprior models.}
\usage{
\method{summary}{iprior}(object, ...)
}
\arguments{
\item{object}{Objects of class \code{iprior}.}
}
\description{
Summary screen for iprior models.
}
\details{
Summary screen for iprior models.
}
\examples{
\donttest{mod <- iprior(Hwt ~ ., data=MASS::cats)}
\donttest{summary(mod)}


}

