% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipriorMod_methods.R
\name{summary.ipriorMod}
\alias{summary.ipriorMod}
\alias{print.ipriorMod}
\title{Print and summary method for I-prior models}
\usage{
\method{print}{ipriorMod}(x, digits = 5, ...)

\method{summary}{ipriorMod}(object, ...)
}
\arguments{
\item{digits}{Number of decimal places for the printed coefficients.}

\item{...}{Not used.}

\item{object, x}{An \code{ipriorMod} object.}
}
\description{
Print and summary method for I-prior models
}
