% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipriorMod_methods.R
\name{update.ipriorMod}
\alias{update.ipriorMod}
\title{Update an I-prior model}
\usage{
\method{update}{ipriorMod}(object, method = NULL, control = list(),
  iter.update = 100, ...)
}
\arguments{
\item{object}{An \code{ipriorMod} object.}

\item{method}{An optional method. See \link[=iprior]{here} for details.}

\item{control}{An optional list of controls for the estimation procedure. See
\link[=iprior]{here} for details.}

\item{iter.update}{The number of additional iterations to update the I-prior
model.}

\item{...}{Not used.}
}
\description{
Update an I-prior model
}
