% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varyinv.R
\name{varyinv}
\alias{varyinv}
\alias{varyinv.ipriorKernel}
\alias{varyinv.ipriorMod}
\title{Extract the variance of the responses}
\usage{
varyinv(object, theta = NULL)

\method{varyinv}{ipriorMod}(object, theta = NULL)

\method{varyinv}{ipriorKernel}(object, theta = NULL)
}
\arguments{
\item{object}{An object of class \code{ipriorMod} or \code{ipriorKernel}.}

\item{theta}{(optional) Evaluates the log-likelihood at \code{theta} which is
of the form \code{theta = c(lambda, psi)}.}
}
\description{
Extract the variance of the responses of an I-prior model.
}
\details{
For \code{ipriorKernel} objects, random values for \code{theta} are used.
}

