% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenCpp}
\alias{eigenCpp}
\title{Eigen decomposition of a matrix in C++.}
\usage{
eigenCpp(X)
}
\arguments{
\item{X}{A symmetric, positive-definite matrix}
}
\description{
Returns the eigenvalues and eigenvectors of a matrix X.
}
\details{
A fast implementation of eigen for symmetric, positive-definite
matrices. This helps speed up the I-prior EM algorithm.
}
