% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastSquare}
\alias{fastSquare}
\title{Multiplying a symmetric matrix by itself in C++.}
\usage{
fastSquare(X)
}
\arguments{
\item{X}{A symmetric matrix}
}
\description{
Returns the square of a symmetric matrix X.
}
\details{
A fast implementation of X^2 for symmetric matrices. This helps
speed up the I-prior EM algorithm.
}
