% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastVDiag}
\alias{fastVDiag}
\title{Computing a quadratic matrix form in C++.}
\usage{
fastVDiag(X, y)
}
\arguments{
\item{X}{A symmetric, square matrix of dimension \code{n} by \code{n}}

\item{y}{A vector of length \code{n}}
}
\description{
Returns XdiagyXT.
}
\details{
A fast implementation of XdiagyXT. This helps speed up
the I-prior EM algorithm.
}
