% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iprior.R
\name{iprior}
\alias{iprior}
\alias{iprior.default}
\alias{iprior.formula}
\alias{iprior.ipriorMod}
\title{Fit an I-prior regression model}
\usage{
\method{iprior}{default}(y, ..., kernel = "linear", method = c("direct",
  "em", "fixed", "canonical", "mixed"), control = list(),
  interactions = NULL, est.lambda = TRUE, est.hurst = FALSE,
  est.lengthscale = FALSE, est.offset = FALSE, est.psi = TRUE,
  fixed.hyp = NULL, lambda = 1, psi = 1, nystrom = FALSE,
  nys.seed = NULL, model = list())

\method{iprior}{formula}(formula, data, kernel = "linear", one.lam = FALSE,
  method = c("direct", "em", "fixed", "canonical", "mixed"),
  control = list(), est.lambda = TRUE, est.hurst = FALSE,
  est.lengthscale = FALSE, est.offset = FALSE, est.psi = TRUE,
  fixed.hyp = NULL, lambda = 1, psi = 1, nystrom = FALSE,
  nys.seed = NULL, model = list(), ...)

\method{iprior}{ipriorMod}(object, method = NULL, control = list(),
  iter.update = 100, ...)
}
\arguments{
\item{y}{Vector of response variables}

\item{...}{Only used when fitting using non-formula, enter the variables
(vectors or matrices) separated by commas.}

\item{kernel}{Character vector indicating the type of kernel for the
variables. Available choices are: \itemize{ \item{\code{"linear"} -
(default) for the linear kernel} \item{\code{"canonical"} - alternative
name for \code{"linear"}} \item{\code{"fbm"}, \code{"fbm,0.5"} - for the
fBm kernel with Hurst coefficient 0.5 (default)} \item{\code{"se"},
\code{"se,1"} - for the SE kernel with lengthscale 1 (default)}
\item{\code{"poly"}, \code{"poly2"}, \code{"poly2,0"} - for the polynomial
kernel of degree 2 with offset 0 (default)} \item{\code{"pearson" - for the
Pearson kernel}}} The \code{kernel} argument can also be a vector of length
equal to the number of variables, therefore it is possible to specify
different kernels for each variables. Note that factor type variables are
assigned the Pearson kernel by default, and that non-factor types can be
forced to use the Pearson kernel (not recommended).}

\item{method}{The estimation method. One of: \itemize{ \item{\code{"direct"}
- for the direct minimisation of the marginal deviance using
\code{optim()}'s L-BFGS method} \item{\code{"em"} - for the EM algorithm}
\item{\code{"mixed"} - combination of the direct and EM methods}
\item{\code{"fixed"} - for just obtaining the posterior regression function
with fixed hyperparameters (default method when setting \code{fixed.hyp =
TRUE})} \item{\code{"canonical"} - an efficient estimation method which
takes advantage of the structure of the linear kernel} }}

\item{control}{1}

\item{interactions}{Character vector to specify the interaction terms. When
using formulas, this is specified automatically, so is not required. Syntax
is \code{"a:b"} to indicate variable \code{a} interacts with variable
\code{b}.}

\item{est.lambda}{Logical. Estimate the scale parameters? Defaults to
\code{TRUE}.}

\item{est.hurst}{Logical. Estimate the Hurst coefficients for fBm kernels?
Defaults to \code{FALSE}.}

\item{est.lengthscale}{Logical. Estimate the lengthscales for SE kernels?
Defaults to \code{FALSE}.}

\item{est.offset}{Logical. Estimate the offsets for polynomial kernels?
Defaults to \code{FALSE}.}

\item{est.psi}{Logical. Estimate the error precision? Defaults to
\code{TRUE}.}

\item{fixed.hyp}{Logical. If \code{TRUE}, then no hyperparameters are
estimated, i.e. all of the above \code{est.x} are set to \code{FALSE}, and
vice versa. If \code{NULL} (default) then all of the \code{est.x} defaults
are respected.}

\item{lambda}{Initial/Default scale parameters. Relevant especially if
\code{est.lambda = FALSE}.}

\item{psi}{Initial/Default value for error precision. Relevant especially if
\code{est.psi = FALSE}.}

\item{nystrom}{Either logical or an integer indicating the number of Nystrom
samples to take. Defaults to \code{FALSE}. If \code{TRUE}, then
approximately 10\% of the sample size is used for the Nystrom
approximation.}

\item{nys.seed}{The random seed for the Nystrom sampling. Defaults to
\code{NULL}, which means the random seed is not fixed.}

\item{model}{DEPRECATED.}

\item{formula}{The formula to fit when using formula interface.}

\item{data}{Data frame containing variables when using formula interface.}

\item{one.lam}{1}

\item{object}{An \code{ipriorKernel2} or \code{ipriorMod} object.}

\item{iter.update}{1}
}
\value{
1
}
\description{
A function to perform regression using I-priors. The I-prior model parameters
may be estimated in a number of ways: direct minimisation of the marginal
deviance, EM algorithm, fixed hyperparameters, or using a Nystrom kernel
approximation.
}
\details{
The \code{iprior()} function is able to take formula based input and
non-formula. When not using formula, the syntax is as per the default S3
method. That is, the response variable is the vector \code{y}, and any
explanatory variables should follow this, and separated by commas.

As described \link[=kernL2]{here}, the model can be loaded first into an
\code{ipriorKernel} object, and then passed to the \code{iprior()} function
to perform the estimation.
}
\section{Methods (by class)}{
\itemize{
\item \code{ipriorMod}: Re-run or continue running the EM algorithm from last
attained parameter values in object \code{ipriorMod}.
}}

\examples{
1

}
