% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iprior.R
\name{iprior}
\alias{iprior}
\alias{iprior.default}
\alias{iprior.formula}
\alias{iprior.ipriorKernel}
\alias{iprior.ipriorMod}
\title{Fit an I-prior regression model}
\usage{
iprior(object, ...)

\method{iprior}{default}(object = NULL, model = list(), control = list(),
  y, ...)

\method{iprior}{formula}(object, data, model = list(), control = list(),
  ...)

\method{iprior}{ipriorKernel}(object, control = list(), ...)

\method{iprior}{ipriorMod}(object, control = list(), ...)
}
\arguments{
\item{object}{This is either an object of class formula (when fitting using
formula interface), \code{ipriorKernel} or \code{ipriorModel}. When left
\code{NULL}, then \code{y} must be used, along with the
covariates/independent variables.}

\item{...}{Only for when fitting using non-formula, enter the variables
(vectors or matrices) separated by commas. No other options applicable
here.}

\item{model}{(optional) List of model options. Not used for
\code{ipriorKernel} or \code{ipriorModel} objects.}

\item{control}{(optional) list of control options for EM algorithm and
output.}

\item{y}{Vector of response variables.}

\item{data}{Data frame containing variables when using formula interface.}
}
\value{
An object of class \code{ipriorMod}.
}
\description{
Description
}
\details{
Details
}
\section{Methods (by class)}{
\itemize{
\item \code{ipriorKernel}: Takes in object of type \code{ipriorKernel} and estimates
the parameters of the model via the EM algorithm.

\item \code{ipriorMod}: Re-run or continue running the EM algorithm from last
attained parameter values in object \code{ipriorMod}.
}}
\examples{
(mod.iprior <- iprior(stack.loss ~ ., data = stackloss))

}

