% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iprior.R
\name{iprior}
\alias{iprior}
\title{Fit an I-prior regression model}
\usage{
iprior(formula, data, model = list(), control = list(), ...)
}
\arguments{
\item{formula}{the model formula to fit}

\item{data}{data frame containing variables}

\item{model}{list of model options}

\item{control}{list of control options for EM algorithm and output}
}
\value{
an object of class iprior
}
\description{
Fit an I-prior regression model
}
\examples{
(mod.iprior <- iprior(stack.loss ~ ., data = stackloss))

}

