% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernel_loader.R
\name{kernL}
\alias{kernL}
\alias{kernL.default}
\title{Load the kernel matrices for \code{iprior} fit}
\usage{
kernL(y, ..., model = list())

\method{kernL}{default}(y, ..., model = list())
}
\arguments{
\item{model}{list of model options}

\item{formula}{the model formula to fit}

\item{data}{data frame containing variables}

\item{control}{list of control options for EM algorithm and output}
}
\value{
an object of class iprior
}
\description{
Load the kernel matrices for \code{iprior} fit
}
\examples{
str(ToothGrowth)
mod <- kernL(y = ToothGrowth$len, supp = ToothGrowth$supp,
             dose = ToothGrowth$dose, model = list(interactions="1:2"))
summary(mod)
\dontrun{mod <- kernL(len ~ supp * dose, data = ToothGrowth)  # also accepts
formula}

}

