% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik}
\alias{logLik}
\alias{logLik.ipriorKernel}
\alias{logLik.ipriorMod}
\title{Function which returns the log-likelihood value of an \code{ipriorMod} or
\code{ipriorKernel} object.}
\usage{
\method{logLik}{ipriorMod}(object, theta = NULL, ...)

\method{logLik}{ipriorKernel}(object, theta = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{ipriorMod} or \code{ipriorKernel}.}

\item{theta}{(optional) evaluates the log-likelihood at \code{theta} which is
of the form \code{theta = c(lambda, psi)}.}
}
\description{
Description.
}
\details{
Details.
}
\examples{
mod.iprior <- iprior(stack.loss ~ ., data=stackloss)
logLik(mod.iprior)

}

