% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{simdat}
\alias{simdat}
\title{Random slopes model simulated data}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{\code{y}}{Response variable}
  \item{\code{x}}{Explanatory variable}
  \item{\code{grp}}{Factor, indicating the group of that observation.}
}}
\usage{
simdat
}
\description{
A simulated dataset to illustrate multilevel modelling using I-priors.
}
\examples{
data(simdat)
str(simdat)
}
\keyword{datasets}

