% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope.R
\name{slope}
\alias{slope}
\title{Function which returns the Variance of Y and log-likelihood value.}
\usage{
slope(object)
}
\arguments{
\item{x}{Object of class "iprior"}

\item{alpha}{(Optional) specify intercept value.}

\item{lambda}{(Optional) specify lambda value.}

\item{psi}{(Optional) specify psi value.}
}
\value{
List of length 2.
}
\description{
\code{slope} returns the slopes.
}
\details{
Test
}
\examples{
mod.iprior <- iprior(stack.loss ~ Air.Flow, data=stackloss)
slope(mod.iprior)
}

